Move out CompileError to core as generic form
